requirements.txt
streamlit>=1.31.0
pydantic>=2.6.1
python-dotenv>=1.0.1
openai>=1.12.0
tavily-python>=0.3.1
langchain>=0.1.9
langchain-community>=0.0.20
rich>=13.7.0
numpy>=1.26.0  # Using a version compatible with Python 3.9+
